+++ service/collector_test.go
@@ -36,8 +36,8 @@ import (
 
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/config/configmapprovider"
+	"go.opentelemetry.io/collector/internal/testcomponents"
 	"go.opentelemetry.io/collector/internal/testutil"
-	"go.opentelemetry.io/collector/service/defaultcomponents"
 )
 
 // TestCollector_StartAsGoRoutine must be the first unit test on the file,
@@ -48,7 +48,7 @@ func TestCollector_StartAsGoRoutine(t *testing.T) {
 	collectorTelemetry = &colTelemetry{}
 	defer func() { collectorTelemetry = preservedAppTelemetry }()
 
-	factories, err := defaultcomponents.Components()
+	factories, err := testcomponents.DefaultFactories()
 	require.NoError(t, err)
 
 	set := CollectorSettings{
@@ -81,7 +81,7 @@ func TestCollector_StartAsGoRoutine(t *testing.T) {
 }
 
 func TestCollector_Start(t *testing.T) {
-	factories, err := defaultcomponents.Components()
+	factories, err := testcomponents.DefaultFactories()
 	require.NoError(t, err)
 
 	loggingHookCalled := false
@@ -151,7 +151,7 @@ func TestCollector_ReportError(t *testing.T) {
 	collectorTelemetry = &mockColTelemetry{}
 	defer func() { collectorTelemetry = preservedAppTelemetry }()
 
-	factories, err := defaultcomponents.Components()
+	factories, err := testcomponents.DefaultFactories()
 	require.NoError(t, err)
 
 	col, err := New(CollectorSettings{